requirements.txt
numpy
pyrtlsdr
pyfftw
pyyaml
pygame_gui
gps